requirements.txt
pandas
tqdm
scikit-learn
tensorboard
pyarrow
torch
pyyaml
polars #polars-lts-cpu
wandb